==> 5.Subclustering.Rmd <==
---
title: "sub_clustering"
author: "Theo"
date: "2024-05-10"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = F)
```

```{r,echo=FALSE, warning=FALSE, message=FALSE}
library(Seurat)
library(patchwork)
set.seed(9999)
library(miloR)
library(scater)
library(tidyverse)
library(viridis)
library(presto)
```

```{r}
object <- readRDS("SCT/intgr_indiv_SCT_harmony/data/annotated_intgr.rds")

DimPlot(object, reduction = "umap", label = TRUE, shuffle = T, group.by = "cell_type")

```

```{r}
T_cell <- subset(object, subset = cell_type== "T")
NK_T <- subset(object, subset  = cell_type== "NK/T")
Ec <-  subset(object, subset  = cell_type== "EC")
Myelo <- subset(object, subset  = cell_type== "M/MC")
Gran <- subset(object, subset  = cell_type== "Gran")
B_cell <- subset(object, subset  = cell_type== "B")
plasm <- subset(object, subset  = cell_type== "plasm/?")
pDCs <- subset(object, subset  = cell_type== "pDCs")
HC1_cell <- subset(object, subset = cell_type == "HEP1/C")
HC2_cell <- subset(object, subset = cell_type == "HEP2/CC")

#rm(object)
```

## T-Cells

```{r}
T_cell <- FindNeighbors(T_cell, reduction = "harmony", dims = 1:50)
T_cell <- FindClusters(T_cell, algorithm = 4, method = "igraph" , resolution = 0.25)
T_cell <- RunUMAP(T_cell, dims = 1:50, assay = "SCT", reduction = "harmony")


# p1 <- DimPlot(T_cell, reduction = "umap", label = TRUE, shuffle = T) + DimPlot(T_cell, reduction = "umap", group.by = "celltype_wide", label = TRUE, shuffle = T)
# p2 <- DimPlot(T_cell, reduction = "umap", label = TRUE, shuffle = T, group.by = "anchor_pred")
# p1


#p2
# ggsave("T_cell_umap_0_25.png", plot = p1, width = 10, height = 10)
# ggsave("T_cell_umap_0.25.pdf", plot = p1, device = "pdf", dpi = 300, width = 10, height = 8)
# p1
# markers
```

```{r}

# T_markers <- wilcoxauc(T_cell, seurat_assay = "SCT","SCT_snn_res.0.25")
# # set filtering params
# top_genes <- top_markers(T_markers, n = 200, auc_min = .5, padj_max = 0.05, pct_in_min = 70)
# write.csv(top_genes, "output/T_cell_Markers_res_0_25.csv", row.names = FALSE)


# pseudo_ifnb <- AggregateExpression(T_cell, assays = "RNA", return.seurat = TRUE, group.by = cell_subtype)
# # each 'cell' is a donor-condition-celltype pseudobulk profile
# head(Cells(pseudo_ifnb))
# Idents(pseudo_ifnb) <- cell_subtype
# 
# bulk.mono.de <- FindMarkers(object = pseudo_ifnb, 
#                          test.use = "DESeq2",cells.1 = pseudo_ifnb@meta.data$)
# head(bulk.mono.de, n = 15)

#saveRDS(T_cell, "T_cell_annot.rds")

```

# Manual annotation of the T cells res 0.25

```{r}
T_manual <- c( "1" = "Exh_CD8_T",
                          "2" = "Naive_T",
                          "3" = "Maf_CD4_T",
                          "4" = "Malat1_T",
                          "5" = "Itgb1_CD8_T",
                          "6" = "Proliferating_T",
                          "7" = "CD3_T",
                          "8" = "NKT_CD8_T"
                          )

T_cell@meta.data$cell_subtype <- T_manual[as.character(T_cell@meta.data$SCT_snn_res.0.25)]

# saveRDS(T_cell, "output/T_cell_annot.rds")


# DimPlot(T_cell, reduction = "umap", label = T, shuffle = T, group.by = "cell_subtype")

#ggsave("output/T_cell_annot_0.25.pdf", plot = T_annot, device = "pdf", dpi = 300, width = 10, height = 8)
```

## Endothelial cells

```{r}
Ec <- FindNeighbors(Ec, reduction = "harmony", dims = 1:50)
Ec <- FindClusters(Ec, algorithm = 4, method = "igraph" , resolution = 0.2)
Ec <- RunUMAP(Ec, dims = 1:50, assay = "SCT", reduction = "harmony")
# p1 <-DimPlot(Ec, reduction = "umap", label = TRUE, shuffle = TRUE, alpha= 0.5) + DimPlot(Ec, reduction = "umap",group.by = "anchor_pred", label = TRUE)
# p2 <-DimPlot(Ec, reduction = "umap", label = TRUE, shuffle = T, group.by = "anchor_pred")
# p1
# ggsave("EC_umap_0_2.pdf", plot = p1, device = "pdf", dpi = 300, width = 10, height = 8)
# ggsave("EC_umap_anchor_pred.pdf", plot = p2, device = "pdf", dpi = 300, width = 10, height = 8)
# 
# # markers
# Ec_markers <- wilcoxauc(Ec, seurat_assay = "SCT","SCT_snn_res.0.2")
# top_genes <- top_markers(Ec_markers, n=200, auc_min = .5, padj_max = 0.05, pct_in_min = 70)
# write.csv(top_genes, "output/Ec_Markers_res_0_2.csv", row.names = FALSE)
# saveRDS(Ec, "Ec.rds")

```

```{r}
EC_manual <- c( "1" = "Homeostatic_EC",
                          "2" = "Oncofetal_IGFBP3+_EC",
                          "3" = "VEGF+_EC",
                          "4" = "Prostaglandin+IL1R+_EC",
                          "5" = "CD52+_EC",
                          "6" = "mixed_EC",
                          "7" = "_EC"
                          )
Ec@meta.data$cell_subtype <- EC_manual[as.character(Ec@meta.data$SCT_snn_res.0.2)]
Ec <-subset(Ec, subset = celltype_wide != "B")# removed Random B cells

DimPlot(Ec, reduction = "umap",group.by = "cell_subtype", label = TRUE)
# 
# saveRDS(Ec, "output/Ec.rds")
# Ec_annot <- DimPlot(Ec, reduction = "umap", label = T, shuffle = T, group.by = "cell_subtype")
# ggsave("output/Ec_cell_annot_0.2.pdf", plot = Ec_annot, device = "pdf", dpi = 300, width = 10, height = 8)


```

## SCT global version doesnt seem to be better

```{r}
# EC_SCT <- SCTransform(EC_SCT, vars.to.regress = "percent.mt", verbose = TRUE, assay = "RNA")
# EC_SCT <- RunPCA(EC_SCT)
# EC_SCT <- FindNeighbors(EC_SCT,  reduction = "harmony", dims = 1:50)
# EC_SCT <- FindClusters(EC_SCT, algorithm = 4, method = "igraph" , resolution = 0.3)
# EC_SCT <- RunUMAP(EC_SCT, dims = 1:50)
# 
# DimPlot(EC_SCT, reduction = "umap", label = TRUE, shuffle = T) + DimPlot(EC_SCT, reduction = "umap", group.by = "condition", label = TRUE, shuffle = T)
# 
# EC_SCT_markers <- wilcoxauc(EC_SCT, seurat_assay = "SCT","SCT_snn_res.0.3")
# # set filtering params
# top_genes <- top_markers(EC_SCT_markers, n = 50, auc_min = .5, padj_max = 0.05, pct_in_min = 70)
# write.csv(top_genes, "EC_SCT_Markers_res_0_3.csv", row.names = FALSE)
# saveRDS(T_cell, "EC_SCT.rds")
```

## Myeloid

```{r}
Myelo <- FindNeighbors(Myelo, reduction = "harmony", dims = 1:50)
Myelo <- FindClusters(Myelo, algorithm = 4, method = "igraph" , resolution = 0.6)
Myelo <- RunUMAP(Myelo, dims = 1:50, assay = "SCT", reduction = "harmony")

# p1 <-DimPlot(Myelo, reduction = "umap", label = TRUE) + DimPlot(Myelo, reduction = "umap",group.by = "anchor_pred", label = TRUE)
# 
# p2 <-DimPlot(Myelo, reduction = "umap", label = TRUE, shuffle = T, group.by = "anchor_pred")
# p1
# ggsave("Myelo_umap_0_6.pdf", plot = p1, device = "pdf", dpi = 300, width = 10, height = 8)
# ggsave("Myelo_umap_anchor_pred.pdf", plot = p2, device = "pdf", dpi = 300, width = 10, height = 8)
# # # markers
# Myelo_markers <- wilcoxauc(Myelo, seurat_assay = "SCT","SCT_snn_res.0.6")
# top_genes <- top_markers(Myelo_markers, n = 200, auc_min = .5, padj_max = 0.05, pct_in_min = 70)
# write.csv(top_genes, "output/Myelo_Markers_res_0_6.csv", row.names = FALSE)
# 
# saveRDS(Myelo, "Myelo.rds")
```

```{r}
Myeloid_manual <- c( "1" = "IL1B+CCR2+_MDMs",
                          "2" = "Stk10_Mono",
                          "3" = "Ifitm3+_Mono",
                          "4" = "CD38+ Apoe+_Mph",
                          "5" = "Cxcl16+Apoe+_Mph",
                          "6" = "cDC2s",
                          "7" = "Prostaglandin+_KCs",
                          "8" = "Malat1_Mono",
                          "9" = "Mertk+_KCs",
                          "10" = "cDC1s",
                          "11" = "Migratory_DCs",
                          "12" = "Prolif_DCs"
                          )
Myelo@meta.data$cell_subtype <- Myeloid_manual[as.character(Myelo@meta.data$SCT_snn_res.0.6)]

# DimPlot(Myelo, reduction = "umap", label = T, shuffle = T, group.by = "cell_subtype")
# 
# saveRDS(Myelo, "output/Myelo_annot.rds")
# 
# 
# 
# ggsave("Myelo_cell_annot_0.6.pdf", plot = myelo_annot, device = "pdf", dpi = 300, width = 10, height = 8)
```

```{r}
Myelo$condition <- factor(Myelo$condition, levels = c('ctl', 'int', 'end'))
vln <-VlnPlot(
    object,
    features=c("Acta2"),
    group.by = "cell_subtype",
    layer = "data",
    split.by = "condition"
    )
#split.by = "condition",
vln
ggsave("Trem2Expression_myeloid.pdf", plot = vln, device = "pdf", dpi = 300, width = 10, height = 8)
```

## Granules

```{r}
Gran <- FindNeighbors(Gran, reduction = "harmony", dims = 1:50)
Gran <- FindClusters(Gran, algorithm = 4, method = "igraph" , resolution = 0.6)
Gran <- RunUMAP(Gran, dims = 1:50, assay = "SCT", reduction = "harmony")
p1 <-DimPlot(Gran, reduction = "umap", label = TRUE) + DimPlot(Gran, reduction = "umap",group.by = "condition", label = TRUE)
p2 <-DimPlot(Gran, reduction = "umap", label = TRUE, shuffle = T, group.by = "anchor_pred")
p1
# ggsave("gran_umap_0_6.pdf", plot = p1, device = "pdf", dpi = 300, width = 10, height = 8)
# # # markers
# Gran_markers <- wilcoxauc(Gran, seurat_assay = "SCT","SCT_snn_res.0.6")
# top_genes <- top_markers(Gran_markers, n = 200, auc_min = .5, padj_max = 0.05, pct_in_min = 70)
# write.csv(top_genes, "output/Gran_Markers_res_0_06.csv", row.names = FALSE)
# saveRDS(Gran, "Gran.rds")
```

```{r}
Gran_manual <- c( "1" = "Cxcr2+Ccl6+_Neu ",
                          "2" = "IL1B+_Neu",
                          "3" = "IL1Ra+PD-L1+_Neu",
                          "4" = "Tox+_Neu",
                          "5" = "Retnlg_Neu",
                          "6" = "Meis2_Neu"
                          )
Gran@meta.data$cell_subtype <- Gran_manual[as.character(Gran@meta.data$SCT_snn_res.0.6)]

# gran_annot <- DimPlot(Gran, reduction = "umap", label = T, shuffle = T, group.by = "cell_subtype")
# ggsave("Gran_cell_annot_0.6.pdf", plot = gran_annot, device = "pdf", dpi = 300, width = 10, height = 8)
```

## Nk/T

```{r}
NK_T <- FindNeighbors(NK_T, reduction = "harmony", dims = 1:50)
NK_T <- FindClusters(NK_T, algorithm = 4, method = "igraph" , resolution = 0.3)
NK_T <- RunUMAP(NK_T, dims = 1:50, assay = "SCT", reduction = "harmony")
p1 <-DimPlot(NK_T, reduction = "umap", label = TRUE) + DimPlot(NK_T, reduction = "umap",group.by = "condition", label = TRUE)
p2 <-DimPlot(NK_T, reduction = "umap", label = TRUE, shuffle = T, group.by = "anchor_pred")
p1

# # markers
# NK_markers <- wilcoxauc(NK_T, seurat_assay = "SCT","SCT_snn_res.0.3")
# top_genes <- top_markers(NK_markers, n = 200, auc_min = .5, padj_max = 0.05, pct_in_min = 70)
# write.csv(top_genes, "output/NK_Markers_res_0_3.csv", row.names = FALSE)
#saveRDS(NK_T, "NK_T.rds")
```

```{r}
NK_manual <- c( "1" = "Cytotoxic/Activated_NK",
                          "2" = "Exhausted_NK",
                          "3" = "Fatty_NK"
                          )
NK_T@meta.data$cell_subtype <- NK_manual[as.character(NK_T@meta.data$SCT_snn_res.0.3)]

DimPlot(NK_T, reduction = "umap", label = T, shuffle = T, group.by = "cell_subtype")
```

## Plasma and other cells

```{r}
plasm <- FindNeighbors(plasm, reduction = "harmony", dims = 1:50)
plasm <- FindClusters(plasm, algorithm = 4, method = "igraph" , resolution = 0.1)
plasm <- RunUMAP(plasm, dims = 1:50, assay = "SCT", reduction = "harmony")
p1 <-DimPlot(plasm, reduction = "umap", label = TRUE) + DimPlot(plasm, reduction = "umap",group.by = "cell_subtype", label = TRUE)
p2 <-DimPlot(plasm, reduction = "umap", label = TRUE, shuffle = T, group.by = "anchor_pred")
p1
# # markers
# plasm_markers <- wilcoxauc(plasm, seurat_assay = "SCT","SCT_snn_res.0.1")
# top_genes <- top_markers(plasm_markers, n = 200, auc_min = .5, padj_max = 0.05, pct_in_min = 70)
# write.csv(top_genes, "output/Plasma_Markers_res_0_1.csv", row.names = FALSE)
#saveRDS(plasm, "plasm.rds")
```

```{r}
plasm_manual <- c( "1" = "Igkc_Plasm",
                          "2" = "Hmgb2_Plasm",
                          "3" = "Plasm_Fib",
                          "4" = "Bnip3l_Plasm"
                          )
plasm@meta.data$cell_subtype <- plasm_manual[as.character(plasm@meta.data$SCT_snn_res.0.1)]
# 
# plasm_annot <- DimPlot(plasm, reduction = "umap", label = T, shuffle = T, group.by = "cell_subtype")
# 
# ggsave("plasm_cell_annot_0.1.pdf", plot = plasm_annot, device = "pdf", dpi = 300, width = 10, height = 8)
```

## PDcs

```{r}
pDCs <- FindNeighbors(pDCs, reduction = "harmony", dims = 1:50)
pDCs <- FindClusters(pDCs, algorithm = 4, method = "igraph" , resolution = 0.13)
pDCs <- RunUMAP(pDCs, dims = 1:50, assay = "SCT", reduction = "harmony")
p1 <-DimPlot(pDCs, reduction = "umap", label = TRUE) + DimPlot(pDCs, reduction = "umap",group.by = "condition", label = TRUE)
p2 <-DimPlot(pDCs, reduction = "umap", label = TRUE, shuffle = T, group.by = "anchor_pred")
p1
# # markers
# pDCs_markers <- wilcoxauc(pDCs, seurat_assay = "SCT","SCT_snn_res.0.13")
# top_genes <- top_markers(pDCs_markers, n = 50, auc_min = .5, padj_max = 0.05, pct_in_min = 70)
# write.csv(top_genes, "output/pDCs_Markers_res_0_13.csv", row.names = FALSE)
#saveRDS(pDCs, "pDCs.rds")
```

```{r}
pDCs_manual <- c( "1" = "Irf8+_pDCs",
           "2" = "Foxp1+_pDCs"
                          )
pDCs@meta.data$cell_subtype <- pDCs_manual[as.character(pDCs@meta.data$SCT_snn_res.0.13)]

# 
# pDCs_annot <- DimPlot(pDCs, reduction = "umap", label = T, shuffle = T, group.by = "cell_subtype")
# 
# ggsave("pDCs_annot_0.13.pdf", plot = pDCs_annot, device = "pdf", dpi = 300, width = 10, height = 8)
```

## B cells

```{r}
B_cell <- FindNeighbors(B_cell, reduction = "harmony", dims = 1:50)
B_cell <- FindClusters(B_cell, algorithm = 4, method = "igraph" , resolution = 0.08)
B_cell <- RunUMAP(B_cell, dims = 1:50, assay = "SCT", reduction = "harmony")

# p1 <-DimPlot(B_cell, reduction = "umap", label = TRUE, shuffle = T) + DimPlot(B_cell, reduction = "umap",group.by = "condition", label = TRUE, shuffle = T)
# p2 <-DimPlot(B_cell, reduction = "umap", label = TRUE, shuffle = T, group.by = "anchor_pred")
# p1


# # markers
# B_markers <- wilcoxauc(B_cell, seurat_assay = "SCT","SCT_snn_res.0.08")
# # set filtering params
# top_genes <- top_markers(B_markers, n = 200, auc_min = .5, padj_max = 0.05, pct_in_min = 70)
# write.csv(top_genes, "output/B_cell_Markers_res_0_08.csv", row.names = FALSE)
#saveRDS(B_cell, "B_cell.rds")
```

```{r}
B_manual <- c( "1" = "Activated_B",
               "2" = "Naive/Early_Ly6D+_B",
               "3" = "Foxp1+Very_early_B"
               )
B_cell@meta.data$cell_subtype <- B_manual[as.character(B_cell@meta.data$SCT_snn_res.0.08)]


# B_annot <- DimPlot(B_cell, reduction = "umap", label = T, shuffle = T, group.by = "cell_subtype")
# 
# ggsave("B_annot_0.08.pdf", plot = B_annot, device = "pdf", dpi = 300, width = 10, height = 8)
```

## hep/ tumor 1

```{r}

HC1_cell <- FindNeighbors(HC1_cell, reduction = "harmony", dims = 1:50)
HC1_cell <- FindClusters(HC1_cell, algorithm = 4, method = "igraph" , resolution = 0.06)
HC1_cell <- RunUMAP(HC1_cell, dims = 1:50, assay = "SCT", reduction = "harmony")

p2 <-DimPlot(HC1_cell, reduction = "umap", label = TRUE) + DimPlot(HC1_cell, reduction = "umap",group.by = "condition", shuffle = T)
p2


# # markers
# HC1_markers <- wilcoxauc(HC1_cell, seurat_assay = "SCT","SCT_snn_res.0.06")
# # set filtering params
# top_genes <- top_markers(HC1_markers, n = 50, auc_min = .5, padj_max = 0.05, pct_in_min = 70)
# write.csv(top_genes, "HC1_markers_0_06.csv", row.names = FALSE)
#saveRDS(HC1_cell, "HC1_cell.rds")

HC1_manual <- c( "1" = "Apoc1+Aldob+_Hep",
                 "2" = "IL6ST+_Hep"
               )
HC1_cell@meta.data$cell_subtype <- HC1_manual[as.character(HC1_cell@meta.data$SCT_snn_res.0.06)]

# DimPlot(HC1_cell, reduction = "umap", label = T, shuffle = T, group.by = "cell_subtype")

```

## hep/ tumor 2

```{r}
HC2_cell <- FindNeighbors(HC2_cell, reduction = "harmony", dims = 1:50)
HC2_cell <- FindClusters(HC2_cell, algorithm = 4, method = "igraph" , resolution = 0.2)
HC2_cell <- RunUMAP(HC2_cell, dims = 1:50, assay = "SCT", reduction = "harmony")
#p1 <- "umap", label = TRUE) + DimPlot(HC2_cell, reduction = "umap",group.by = "condition", shuffle = T)
DimPlot(HC2_cell, reduction = "umap", label = TRUE) 


# # # markers
# HC2_markers <- wilcoxauc(HC2_cell, seurat_assay = "SCT","SCT_snn_res.0.15")
# # # set filtering params
# top_genes <- top_markers(HC2_markers, n = 50, auc_min = .5, padj_max = 0.05, pct_in_min = 70)
# write.csv(top_genes, "HC2_markers_0_15.csv", row.names = FALSE)
#saveRDS(HC2_cell, "HC2_cell.rds")
HC2_manual <- c( "1" = "Vegfa+_Hep",
               "2" = "CD74+MHCII+_Hep",
               "3" = "Serpina+_Hep"
               )
HC2_cell@meta.data$cell_subtype <- HC2_manual[as.character(HC2_cell@meta.data$SCT_snn_res.0.2)]

# DimPlot(HC2_cell, reduction = "umap", label = T, shuffle = T, group.by = "cell_subtype")

```
### After Succesfull annotation we put the metdata information to the original object (i removed B cells from EC clustere and they are NA(cell_subtype) in the integrated dataset) But also put on the wide clustering Myeloid differentiation and Fibroblasts

```{r}


meta1 <- T_cell@meta.data[, "cell_subtype", drop = FALSE]
meta2 <- NK_T@meta.data[, "cell_subtype", drop = FALSE]
meta3 <- Ec@meta.data[, "cell_subtype", drop = FALSE]
meta4 <- Myelo@meta.data[, "cell_subtype", drop = FALSE]
meta5 <- Gran@meta.data[, "cell_subtype", drop = FALSE]
meta6 <- B_cell@meta.data[, "cell_subtype", drop = FALSE]
meta7 <- plasm@meta.data[, "cell_subtype", drop = FALSE]
meta8 <- pDCs@meta.data[, "cell_subtype", drop = FALSE]
meta9 <- HC1_cell@meta.data[, "cell_subtype", drop = FALSE]
meta10 <- HC2_cell@meta.data[, "cell_subtype", drop = FALSE]

combined_meta <- rbind(meta1, meta2, meta3, meta4, meta5, meta6, meta7, meta8, meta9 , meta10)

object <- AddMetaData(object, metadata = combined_meta, col.name = "cell_subtype")
DefaultAssay(object) <- "SCT"
Idents(object = object) <- "cell_subtypes"
#object <- factor(object$condition, levels = c('ctl', 'int', 'end'))


metadata <- object@meta.data
# Seperate mph dc and others in the bigger subcluster 
metadata$cell_type <- ifelse(grepl("_Mph$", metadata$cell_subtype), "Mph",
                      ifelse(grepl("_DCs$", metadata$cell_subtype), "DCs",
                      ifelse(grepl("_KCs$", metadata$cell_subtype), "KCs",
                      ifelse(grepl("_MDMs$", metadata$cell_subtype), "MDMs",
                      ifelse(grepl("_Mono$", metadata$cell_subtype), "Mono",
                      ifelse(grepl("cDC1s$", metadata$cell_subtype), "DCs",
                      ifelse(grepl("cDC2s$", metadata$cell_subtype), "DCs",
                      ifelse(grepl("_Fib$", metadata$cell_subtype), "Fib",
                      metadata$cell_type))))))))

                      
object@meta.data <- metadata
unique(object$cell_type)






saveRDS(object, file= "SCT/intgr_indiv_SCT_harmony/Subclustering/subcluster_annotated_RasV.rds")
```

```{r}
annot_umap <- DimPlot(object, reduction = "umap", label = T, shuffle = T, group.by = "cell_subtype")
annot_umap
ggsave("Subcluster_annotated_RasV.pdf", plot = annot_umap, device = "pdf", dpi = 300, width = 16, height = 8)
```
# Enrichment based on the dataset from oncofetal

## First transform human markers from the paper to mouse ones, the xlsx file has 2 layers one from myeloid and one for endothelial cells

```{r}

#devtools::install_github("montilab/hypeR")
# install.packages("readxl")   # If the table is in an Excel file
# install.packages("openxlsx") # If the table is in an Excel file
# install.packages("tidyverse") # For data manipulation
# if (!requireNamespace("BiocManager", quietly = TRUE))
#     install.packages("BiocManager")
# BiocManager::install("biomaRt")
# 
# if (!requireNamespace("BiocManager", quietly = TRUE))
#     install.packages("BiocManager")
# BiocManager::install("biomaRt")
library(readxl)
library(openxlsx)
library(tidyverse)
library(biomaRt)
library(anndataR)
library(hypeR)
library(biomaRt)
```

```{r}


path <- ("SCT/intgr_indiv_SCT_harmony/Subclustering/markers_enrichment_oncofetal/mmc.xlsx")

MNC_df <- read_excel(path, sheet = 1)
EC_df <- read_excel(path, sheet = 2)

# based on how my  the excel is structured
MNC_df <- MNC_df[-1,-1]
new_colnames <- as.character(MNC_df[1, ])
MNC_df <- MNC_df[-1, ]
colnames(MNC_df) <- new_colnames
EC_df <- EC_df[-1,-1]
new_colnames <- as.character(EC_df[1, ])
EC_df <- EC_df[-1, ]
colnames(EC_df) <- new_colnames
```

```{r}
# put the human gene names from the dataframe in a vector to input LDS
EC_human_genes <- unique(as.vector(as.matrix(EC_df)))
MNC_human_genes <- unique(as.vector(as.matrix(MNC_df)))

#listEnsemblArchives()
#listMarts()

# human_mart <- useEnsembl(biomart = "ensembl", dataset = "hsapiens_gene_ensembl", , host = "https://jul2022.archive.ensembl.org",verbose = T, mirror = "www")
# mouse_mart <- useEnsembl(biomart = "ensembl", dataset = "mmusculus_gene_ensembl", , host = "https://jul2022.archive.ensembl.org", verbose = T,mirror = "www")

human <- useMart("ensembl", dataset = "hsapiens_gene_ensembl", host = "https://dec2021.archive.ensembl.org/") # human
mart <- useMart("ensembl", dataset = "mmusculus_gene_ensembl", host = "https://dec2021.archive.ensembl.org/") #mouse only this version of december works

## Endothelial clusters
# Get the orthologous mouse genes using human gene symbols
EC_mapping <- getLDS(
  attributes = c("hgnc_symbol"), # Use human gene symbols as the attribute
  filters = "hgnc_symbol", # Use human gene symbols as the filter
  values = EC_human_genes, # Your list of human gene symbols
  mart = human, # Human mart
  attributesL = c("mgi_symbol"), # Retrieve mouse gene symbols
  martL = mart, # Mouse mart
  uniqueRows = TRUE
)

#monocytes/macrophages clusters
MNC_mapping <- getLDS(
  attributes = c("hgnc_symbol"), 
  filters = "hgnc_symbol", 
  values = MNC_human_genes, 
  mart = human, 
  attributesL = c("mgi_symbol"), 
  martL = mart, 
  uniqueRows = TRUE
)

colnames(EC_mapping) <- c("human_gene", "mouse_gene")
colnames(MNC_mapping) <- c("human_gene", "mouse_gene")
```

# Function to replace the gene names

```{r}
replace_with_mouse <- function(human_gene_name, mapping_df) {
  mouse_genes <- mapping_df$mouse_gene[mapping_df$human_gene == human_gene_name]
  
  if(length(mouse_genes) == 0) {
    return(human_gene_name)  # Return NA if no mouse gene is found or we can use the Human gene names / or we later input the ones manualy based on the new db
  } else {
    return(paste(mouse_genes, collapse = ", "))
  }
}
```

## Some gene names dont have any orthologues or they do but the version doesnt include it

```{r}
EC_df[] <- lapply(EC_df, function(column) sapply(column, replace_with_mouse, mapping_df = EC_mapping))
MNC_df[] <- lapply(MNC_df, function(column) sapply(column, replace_with_mouse, mapping_df = MNC_mapping))


# Check the updated data frame
print(EC_df)
# write.csv(EC_df, "EC_mouse.csv", row.names = FALSE)
# write.csv(MNC_df, "MNC_mouse.csv", row.names = FALSE)
```

## Enrichment of EC based on oncofetal paper

```{r}
# Import my endothelial cell DGE markers
EC_HCC <- read_csv("Ec_Markers_res_0_2.csv")
EC_HCC <- EC_HCC %>% dplyr::select(-rank) # remove the rank column

### reorder the clusters based on numeric values in the colnames
colnames_numeric <- as.numeric(gsub("\\D", "", colnames(EC_HCC)))
# order the column names based on the numeric part
ordered_colnames <- colnames(EC_HCC)[order(colnames_numeric)]
# reorder
EC_HCC <- EC_HCC[, ordered_colnames]

new_colnames <- paste0("cl", seq_along(colnames(EC_HCC))) #instead of 1 2 3 put cl1,cl2/./
colnames(EC_HCC) <- new_colnames # my non annotated subclusters  # my non annotated subclusters
EC_HCC_list <- lapply(EC_HCC, as.vector) # transform the df into a list of vectors (each subcluster)


colnames(EC_df) <- c("0_PLPP3+EC", "1_PLPP3+EC", "2_CD9+EC", "3_PLVAP+EC", "4_PLVAP+EC", "5-CD9+EC",	"6_CD320EC", "7_IGFBP3+EC", "8_IGFBP3+EC", "9_IGFBP3+EC", "10_TFF3+EC")

signatures_list <- lapply(EC_df, as.vector)

enrich_results <- hypeR(signatures_list, EC_HCC_list, test="hypergeometric", background=23467)
hyp_to_excel(enrich_results, file_path="enrich_results_EC_300markers.xlsx")

```

## ####Use hyper R for enrichment analysis

```{r}
file_path <- "enrich_results_EC_300markers.xlsx"


sheets <- excel_sheets(file_path)

combined_df <- lapply(sheets, function(sheet) {
  df <- read_excel(file_path, sheet = sheet)
  df$sheet_name <- sheet
  df
}) %>% bind_rows()


combined_df$fdr <- as.numeric(combined_df$fdr)

# FDR < 0.05
filtered_df <- combined_df %>% filter(fdr < 0.05)

# curate the dataframe for plotting
plot_df <- filtered_df %>% 
  mutate(overlap_size = as.numeric(overlap)) %>% # change the 24 number based on the amount of signature
  replace_na(list(overlap_size = 0))

# PLOT the results
EC_enrich <- ggplot(plot_df, aes(x = sheet_name, y = label)) +
              geom_point(aes(size = overlap_size, color = fdr)) +
              scale_size_continuous(range = c(2, 10)) +
              scale_color_viridis_c(option = "magma", direction = -1) +
              theme_minimal(base_family = "sans") +
              theme(
                plot.background = element_rect(fill = "white", color = NA),
                panel.background = element_rect(fill = "grey60", color = NA),  
                panel.grid.major = element_line(color = "grey100"),
                panel.grid.minor = element_line(color = "grey80"),
                axis.text.x = element_text(angle = 45, hjust = 1, color = "black"),
                axis.text.y = element_text(color = "black"),
                axis.title = element_text(color = "black"),
                legend.background = element_rect(fill = "grey90", color = NA),
                legend.key = element_rect(fill = "grey90", color = NA),
                legend.title = element_text(size = 10, color = "black"),
                legend.text = element_text(size = 10, color = "black"),
                legend.key.size = unit(0.5, "cm"),  # Adjusted legend key size
                plot.title = element_text(hjust = 0.5, size = 14, face = "bold", color = "black")  # Center the title
              ) +
              labs(
                title = "Endothelial Cell Enrichment",
                x = "Signature",
                y = "Label",
                size = "Overlap",
                color = "FDR (<0.05)"
              ) +
              theme(
                plot.title = element_text(hjust = 0.5, size = 14, face = "bold"),
                axis.text.x = element_text(angle = 45, hjust = 1)
  )
EC_enrich
ggsave("EC_dotplot_Enrichment.pdf", plot = EC_enrich, device = "pdf", dpi = 300, width = 10, height = 7, units = "in")
```

## T-CELL enrichment

Import the signature datasets

```{r}
# first T_cell signatures
path1 <- ("SCT/intgr_indiv_SCT_harmony/Subclustering/T_cell_markers_enrichment/Tcellsignatures.xlsx")
#Second signatures (should we combine them?)
path2 <- ("SCT/intgr_indiv_SCT_harmony/Subclustering/T_cell_markers_enrichment/Tcellsignatures_Zheng_Zhang_summarised.xlsx")

#LIPID ASSOCIATED MACROPHAGES-HUMAN
path3 <- ("SCT/intgr_indiv_SCT_harmony/Subclustering/LAMS_markers_enrichment/mmc6.xlsx")

T_cell_sign <- read_excel(path1)
T_cell_sign_zheng <- read_excel(path2)


lams <- read_excel(path3)
# filter from l2fc 1.3 and up (includes all the main signatures), from human fc lams >= 1.3
# lams <- lams %>%
#   filter(log2fcMac3vsMac1>=1.3)%>%
#   filter(pval < 0.05)
lams <- lams %>%
  filter(fcLAMvsMacHuman>=1.3)
        
  
        
lams <- unique(as.vector(lams$gene))
        
```

Map the human to mouse cells

```{r}
Tcell_1 <- unique(as.vector(as.matrix(T_cell_sign)))
Tcell_2 <- unique(as.vector(as.matrix(T_cell_sign_zheng)))

human <- useMart("ensembl", dataset = "hsapiens_gene_ensembl", host = "https://dec2021.archive.ensembl.org/") # human
mouse <- useMart("ensembl", dataset = "mmusculus_gene_ensembl", host = "https://dec2021.archive.ensembl.org/")
```

```{r}
Tcell_1_mapping <- getLDS(
  attributes = c("hgnc_symbol"), # Use human gene symbols as the attribute
  filters = "hgnc_symbol", # Use human gene symbols as the filter
  values = Tcell_1, # Your list of human gene symbols
  mart = human, # Human mart
  attributesL = c("mgi_symbol"), # Retrieve mouse gene symbols
  martL = mouse, # Mouse mart
  uniqueRows = TRUE
)

Tcell_2_mapping <- getLDS(
  attributes = c("hgnc_symbol"), # Use human gene symbols as the attribute
  filters = "hgnc_symbol", # Use human gene symbols as the filter
  values = Tcell_2, # Your list of human gene symbols
  mart = human, # Human mart
  attributesL = c("mgi_symbol"), # Retrieve mouse gene symbols
  martL = mouse, # Mouse mart
  uniqueRows = TRUE
)

Lams_mapping <- getLDS(
  attributes = c("hgnc_symbol"), # Use human gene symbols as the attribute
  filters = "hgnc_symbol", # Use human gene symbols as the filter
  values = lams, # Your list of human gene symbols
  mart = human, # Human mart
  attributesL = c("mgi_symbol"), # Retrieve mouse gene symbols
  martL = mouse, # Mouse mart
  uniqueRows = TRUE
)


colnames(Tcell_1_mapping) <- c("human_gene", "mouse_gene")
colnames(Tcell_2_mapping) <- c("human_gene", "mouse_gene")

colnames(Lams_mapping) <- c("human_gene", "mouse_gene")
```

## Here i call the function i made in the previous EC enrichment

Run first the function : replace_with_mouse

```{r}
T_cell_sign[] <- lapply(T_cell_sign, function(column) sapply(column, replace_with_mouse, mapping_df = Tcell_1_mapping))

T_cell_sign_zheng[] <- lapply(T_cell_sign_zheng, function(column) sapply(column, replace_with_mouse, mapping_df = Tcell_2_mapping))

lams <- Lams_mapping$mouse_gene


# Check the updated data frame
print(lams)
```

## HypeR and plots

```{r}
Tcell_HCC <- read_csv("T_cell_450Markers_res_0_25.csv")
Tcell_HCC <- Tcell_HCC %>% dplyr::select(-rank) # remove the rank column !!!!



### reorder the clusters based on numeric values in the colnames
colnames_numeric <- as.numeric(gsub("\\D", "", colnames(Tcell_HCC)))

# order the column names based on the numeric part
ordered_colnames <- colnames(Tcell_HCC)[order(colnames_numeric)]

# reorder
Tcell_HCC <- Tcell_HCC[, ordered_colnames]

new_colnames <- paste0("cl", seq_along(colnames(Tcell_HCC))) #instead of 1 2 3 put cl1,cl2/./
colnames(Tcell_HCC) <- new_colnames # my non annotated subclusters


Tcell_HCC <- lapply(Tcell_HCC, as.vector) # transform the df into a list of vectors (each subcluster)



signatures_list_1 <- lapply(T_cell_sign, as.vector)
signatures_list_2_zheng <- lapply(T_cell_sign_zheng, as.vector)

##hypeR
T_cell_enrich_results <- hypeR(signatures_list_1, Tcell_HCC, test="hypergeometric", background=23467)

T_cell_zheng_enrich_results <- hypeR(signatures_list_2_zheng, Tcell_HCC, test="hypergeometric", background=23467)

hyp_to_excel(T_cell_enrich_results, file_path="enrich_results_T_300markers.xlsx")
hyp_to_excel(T_cell_zheng_enrich_results, file_path="enrich_results_T_Zheng_450markers_0_25res.xlsx")

```

```{r}
myeloid_HCC <- read_csv("Myelo_Markers_res_0_6.csv")
myeloid_HCC <- myeloid_HCC %>% dplyr::select(-rank) # remove the rank column !!!!


myeloid_hcc <- lapply(myeloid_HCC, as.vector) # transform the df into a list of vectors (each subcluster)


##hypeR
Myeloid_cell_enrich_results <- hypeR(lams, myeloid_hcc, test="hypergeometric", background=23467)



hyp_to_excel(Myeloid_cell_enrich_results, file_path="enrich_results_Myeloid_200markers.xlsx")
file_path <- "enrich_results_Myeloid_200markers.xlsx"
sheets <- excel_sheets(file_path)

combined_df <- lapply(sheets, function(sheet) {
  df <- read_excel(file_path, sheet = sheet)
  df$sheet_name <- sheet
  df
}) %>% bind_rows()
# FDR < 0.05
filtered_df <- combined_df %>% filter(fdr < 0.05)
plot_df <- filtered_df %>% 
  mutate(overlap_size = as.numeric(overlap)) %>% # change the 24 number based on the amount of signature
  replace_na(list(overlap_size = 0))

# PLOT the results
myeloid_enrich <- ggplot(plot_df, aes(x = sheet_name, y = label)) +
              geom_point(aes(size = overlap_size, color = fdr)) +
              scale_size_continuous(range = c(2, 10)) +
              scale_color_viridis_c(option = "magma", direction = -1) +
              theme_minimal(base_family = "sans") +
              theme(
                plot.background = element_rect(fill = "white", color = NA),
                panel.background = element_rect(fill = "grey60", color = NA),  
                panel.grid.major = element_line(color = "grey100"),
                panel.grid.minor = element_line(color = "grey80"),
                axis.text.x = element_text(angle = 45, hjust = 1, color = "black"),
                axis.text.y = element_text(color = "black"),
                axis.title = element_text(color = "black"),
                legend.background = element_rect(fill = "grey90", color = NA),
                legend.key = element_rect(fill = "grey90", color = NA),
                legend.title = element_text(size = 10, color = "black"),
                legend.text = element_text(size = 10, color = "black"),
                legend.key.size = unit(0.5, "cm"),  # Adjusted legend key size
                plot.title = element_text(hjust = 0.5, size = 14, face = "bold", color = "black")  # Center the title
              ) +
              labs(
                title = "Human LAMs Enrichment",
                x = "Signature",
                y = "Label",
                size = "Overlap",
                color = "FDR (<0.05)"
              ) +
              theme(
                plot.title = element_text(hjust = 0.5, size = 14, face = "bold"),
                axis.text.x = element_text(angle = 45, hjust = 1)
  )
myeloid_enrich
ggsave("Human_Lams_Enrichment.pdf", plot = myeloid_enrich, device = "pdf", dpi = 300, width = 10, height = 7, units = "in")
```

```{r}

file_path <- "enrich_results_T_Zheng_450markers_0_25res.xlsx"


sheets <- excel_sheets(file_path)

combined_df <- lapply(sheets, function(sheet) {
  df <- read_excel(file_path, sheet = sheet)
  df$sheet_name <- sheet
  df
}) %>% bind_rows()


combined_df$fdr <- as.numeric(combined_df$fdr)

# FDR < 0.05
filtered_df <- combined_df %>% filter(fdr < 0.05)

# curate the dataframe for plotting
plot_df <- filtered_df %>% 
  mutate(overlap_size = as.numeric(overlap)) %>% # change the 24 number based on the amount of signature
  replace_na(list(overlap_size = 0))

# PLOT the results
T_enrich <- ggplot(plot_df, aes(x = sheet_name, y = label)) +
              geom_point(aes(size = overlap_size, color = fdr)) +
              scale_size_continuous(range = c(2, 10)) +
              scale_color_viridis_c(option = "magma", direction = -1) +
              theme_minimal(base_family = "sans") +
              theme(
                plot.background = element_rect(fill = "white", color = NA),
                panel.background = element_rect(fill = "grey60", color = NA),  
                panel.grid.major = element_line(color = "grey100"),
                panel.grid.minor = element_line(color = "grey80"),
                axis.text.x = element_text(angle = 45, hjust = 1, color = "black"),
                axis.text.y = element_text(color = "black"),
                axis.title = element_text(color = "black"),
                legend.background = element_rect(fill = "grey90", color = NA),
                legend.key = element_rect(fill = "grey90", color = NA),
                legend.title = element_text(size = 10, color = "black"),
                legend.text = element_text(size = 10, color = "black"),
                legend.key.size = unit(0.5, "cm"),  # Adjusted legend key size
                plot.title = element_text(hjust = 0.5, size = 14, face = "bold", color = "black")  # Center the title
              ) +
              labs(
                title = "T Cell Enrichment",
                x = "Signature",
                y = "Label",
                size = "Overlap",
                color = "FDR (<0.05)"
              ) +
              theme(
                plot.title = element_text(hjust = 0.5, size = 14, face = "bold"),
                axis.text.x = element_text(angle = 45, hjust = 1)
  )
T_enrich
ggsave("T_dotplot_Enrichment_1.pdf", plot = T_enrich, device = "pdf", dpi = 300, width = 10, height = 7, units = "in")
```

# Generate the plot from 2nd T_cell signatures

```{r}
file_path <- "enrich_results_T_Zheng_300markers.xlsx"


sheets <- excel_sheets(file_path)

combined_df <- lapply(sheets, function(sheet) {
  df <- read_excel(file_path, sheet = sheet)
  df$sheet_name <- sheet
  df
}) %>% bind_rows()


combined_df$fdr <- as.numeric(combined_df$fdr)

# FDR < 0.05
filtered_df <- combined_df %>% filter(fdr < 0.05)

# curate the dataframe for plotting
plot_df <- filtered_df %>% 
  mutate(overlap_size = as.numeric(overlap)) %>% # change the 24 number based on the amount of signature
  replace_na(list(overlap_size = 0))

# PLOT the results
T_enrich_2 <- ggplot(plot_df, aes(x = sheet_name, y = label)) +
              geom_point(aes(size = overlap_size, color = fdr)) +
              scale_size_continuous(range = c(2, 10)) +
              scale_color_viridis_c(option = "magma", direction = -1) +
              theme_minimal(base_family = "sans") +
              theme(
                plot.background = element_rect(fill = "white", color = NA),
                panel.background = element_rect(fill = "grey60", color = NA),  
                panel.grid.major = element_line(color = "grey100"),
                panel.grid.minor = element_line(color = "grey80"),
                axis.text.x = element_text(angle = 45, hjust = 1, color = "black"),
                axis.text.y = element_text(color = "black"),
                axis.title = element_text(color = "black"),
                legend.background = element_rect(fill = "grey90", color = NA),
                legend.key = element_rect(fill = "grey90", color = NA),
                legend.title = element_text(size = 10, color = "black"),
                legend.text = element_text(size = 10, color = "black"),
                legend.key.size = unit(0.5, "cm"),  # Adjusted legend key size
                plot.title = element_text(hjust = 0.5, size = 14, face = "bold", color = "black")  # Center the title
              ) +
              labs(
                title = "T Cell Enrichment",
                x = "Signature",
                y = "Label",
                size = "Overlap",
                color = "FDR (<0.05)"
              ) +
              theme(
                plot.title = element_text(hjust = 0.5, size = 14, face = "bold"),
                axis.text.x = element_text(angle = 45, hjust = 1)
  )
T_enrich_2
ggsave("T_dotplot_Enrichment_2_Zheng.pdf", plot = T_enrich_2, device = "pdf", dpi = 300, width = 10, height = 7, units = "in")


```

```{r}
# Default
hyp_dots(T_cell_enrich_results, fdr = 0.05, size_by = "significance", merge = F)
```



```{r}
metadata <- T_cell@meta.data
# percentage of cells for each subcluster, condition, and batch/repl
percentage_data <- metadata %>%
  group_by(cell_subtype, condition) %>%
  summarize(count = n(), .groups = 'drop') %>%
  group_by(cell_subtype) %>%
  mutate(total_count = sum(count)) %>%
  ungroup() %>%
  mutate(percentage = (count / total_count) * 100)


# reorder for plotting
condition_order <- c("ctl", "int", "end")
percentage_data$condition <- factor(percentage_data$condition, levels = condition_order)
custom_colors <- c("end" = "#00BA38", "int" = "#619CFF", "ctl" = "#F8766D")#F8766D

condition_order <- c("ctl", "int", "end")

# Reorder conditions in percentage_data
percentage_data$condition <- factor(percentage_data$condition, levels = condition_order)

plots <- percentage_data %>%
  ggplot(aes(x = condition, y = percentage, fill = condition)) +
  geom_bar(stat = "identity", position = position_dodge(width = 0.9)) +
  facet_wrap(~ cell_subtype) +
  theme_minimal() +
  labs(title = "Percentage of Cells by Condition for B cell Subtypes",
       x = "Condition",
       y = "Mean Percentage of Cells ± SD",
       fill = "Condition") +
  scale_y_continuous(labels = scales::percent_format(scale = 1))+
  theme(plot.title = element_text(hjust = 0.5))


# Print the plot
print(plots)
ggsave("B_cell_subcluster_barplot.pdf", plot = plots, device = "pdf", dpi = 300, width = 10, height = 8)
```

## Box plot for better representation of the abudances

```{r}
metadata <- NK_T@meta.data
# percentage of cells for each subcluster, condition, and batch/repl
percentage_data <- metadata %>%
  group_by(cell_subtype, condition, batch) %>%
  summarize(count = n(), .groups = 'drop') %>%
  group_by(cell_subtype) %>%
  mutate(total_count = sum(count)) %>%
  ungroup() %>%
  mutate(percentage = (count / total_count) * 100)


# reorder for plotting
condition_order <- c("ctl", "int", "end")
percentage_data$condition <- factor(percentage_data$condition, levels = condition_order)
custom_colors <- c("end" = "#00BA38", "int" = "#619CFF", "ctl" = "#F8766D")#F8766D

plots <- percentage_data %>%
  ggplot(aes(x = condition, y = percentage, fill = condition)) +
  geom_boxplot(position = position_dodge(width = 0.9), alpha = 0.7) +
  geom_jitter(width = 0.2, alpha = 0.5, size = 0.5) +
  stat_summary(fun = mean, geom = "point", shape = 21, size = 2, color = "black", fill = "white") +
  facet_wrap(~ cell_subtype) +
  theme_minimal() +
  theme(
    plot.title = element_text(hjust = 0.5, face = "bold", size = 14),
    axis.title = element_text(face = "bold", size = 12),
    axis.text = element_text(size = 10),
    strip.text = element_text(face = "bold", size = 12),
    legend.position = "none"
  ) +
  labs(title = "Percentage of Cells by Condition for NK cells Subtypes",
       x = "Condition",
       y = "Percentage of Cells") +
  scale_y_continuous(labels = scales::percent_format(scale = 1))+
  scale_fill_manual(values = custom_colors) #  custom colors




print(plots)
ggsave("Nk_cell_subcluster_boxplot.pdf", plot = plots, device = "pdf", dpi = 300, width = 10, height = 8)

```

# MiloR in subclusters

```{r}
Myelo <- readRDS("SCT/intgr_indiv_SCT_harmony/Subclustering/2.Myeloid/Myelo_annot.rds")
T_cell <- readRDS("SCT/intgr_indiv_SCT_harmony/Subclustering/3.T_cells/T_cell_annot.rds")
Ec <- readRDS("SCT/intgr_indiv_SCT_harmony/Subclustering/1.EC/Ec.rds")


DefaultAssay(object= Myelo)<- "SCT"
DefaultAssay(object= T_cell)<- "SCT"
DefaultAssay(object= Ec)<- "SCT"




T_cell <- as.SingleCellExperiment(T_cell)
Ec <- as.SingleCellExperiment(Ec)
Myelo <- as.SingleCellExperiment(Myelo)


T_milo <- Milo(T_cell) #turn to milo obj
Ec_milo <- Milo(Ec)
Myelo_milo <- Milo(Myelo)

T_milo <- buildGraph(T_milo, k = 40, d = 50, reduced.dim = "HARMONY")
T_milo <- makeNhoods(T_milo, prop = 0.4, k = 40, d=50, refined = TRUE)
plotNhoodSizeHist(T_milo)

Ec_milo <- buildGraph(Ec_milo, k = 35, d = 50, reduced.dim = "HARMONY")
Ec_milo <- makeNhoods(Ec_milo, prop = 0.4, k = 35, d=50, refined = TRUE)
plotNhoodSizeHist(Ec_milo)

Myelo_milo <- buildGraph(Myelo_milo, k = 35, d = 50, reduced.dim = "HARMONY")
Myelo_milo <- makeNhoods(Myelo_milo, prop = 0.4, k =35, d=50, refined = TRUE)
plotNhoodSizeHist(Myelo_milo)

```

```{r}
T_milo <- countCells(T_milo, meta.data = data.frame(colData(T_cell)), sample="batch")
Ec_milo <- countCells(Ec_milo, meta.data = data.frame(colData(Ec)), sample="batch")
Myelo_milo <- countCells(Myelo_milo, meta.data = data.frame(colData(Myelo_milo)), sample="batch")
```

```{r}

T_design <- data.frame(colData(T_milo))[,c("batch", "condition")]
T_design <- distinct(T_design)
rownames(T_design) <- T_design$batch

## reorder rownames to match columns of nhoodCounts(milo)
T_design <- T_design[colnames(nhoodCounts(T_milo)), , drop=FALSE]

table(T_design$condition)


Ec_design <- data.frame(colData(Ec_milo))[,c("batch", "condition")]
Ec_design <- distinct(Ec_design)
rownames(Ec_design) <- Ec_design$batch

## reorder rownames to match columns of nhoodCounts(milo)
Ec_design <- Ec_design[colnames(nhoodCounts(Ec_milo)), , drop=FALSE]

table(Ec_design$condition)

Myelo_design <- data.frame(colData(Myelo_milo))[,c("batch", "condition")]
Myelo_design <- distinct(Myelo_design)
rownames(Myelo_design) <- Myelo_design$batch

## reorder rownames to match columns of nhoodCounts(milo)
Myelo_design <- Myelo_design[colnames(nhoodCounts(Myelo_milo)), , drop=FALSE]

table(Myelo_design$condition)
```

```{r}
T_design_ctl_int <- T_design[T_design$condition %in% c("ctl", "int"), ]
T_design_int_end <- T_design[T_design$condition %in% c("int", "end"), ]
T_design_int_end$condition <- factor(T_design_int_end$condition, levels = c('int', 'end'))


Ec_design_ctl_int <- Ec_design[Ec_design$condition %in% c("ctl", "int"), ]
Ec_design_int_end <- Ec_design[Ec_design$condition %in% c("int", "end"), ]
Ec_design_int_end$condition <- factor(Ec_design_int_end$condition, levels = c('int', 'end'))

Myelo_design_ctl_int <- Myelo_design[Myelo_design$condition %in% c("ctl", "int"), ]
Myelo_design_int_end <- Myelo_design[Myelo_design$condition %in% c("int", "end"), ]
Myelo_design_int_end$condition <- factor(Myelo_design_int_end$condition, levels = c('int', 'end'))
```

```{r}
T_milo <- calcNhoodDistance(T_milo, d=50, reduced.dim = "HARMONY", use.assay = "logcounts") # use assay
da_T_ctl_int <- testNhoods(T_milo, design = ~ condition, design.df = T_design_ctl_int, fdr.weighting="graph-overlap")
da_T_int_end <- testNhoods(T_milo, design= ~ condition, design.df = T_design_int_end, fdr.weighting="graph-overlap")


Ec_milo <- calcNhoodDistance(Ec_milo, d=50, reduced.dim = "HARMONY", use.assay = "logcounts") # use assay
da_Ec_ctl_int <- testNhoods(Ec_milo, design = ~ condition, design.df = Ec_design_ctl_int, fdr.weighting="graph-overlap")
da_Ec_int_end <- testNhoods(Ec_milo, design = ~  condition, design.df = Ec_design_int_end, fdr.weighting="graph-overlap")

Myelo_milo <- calcNhoodDistance(Myelo_milo, d=50, reduced.dim = "HARMONY", use.assay = "logcounts") # use assay
da_Myelo_ctl_int <- testNhoods(Myelo_milo, design = ~ condition, design.df = Myelo_design_ctl_int, fdr.weighting="graph-overlap")
da_Myelo_int_end <- testNhoods(Myelo_milo, design = ~  condition, design.df = Myelo_design_int_end, fdr.weighting="graph-overlap")
```

\

```{r}
da_T_ctl_int %>%
  arrange(- SpatialFDR) %>%
  head() 
da_Ec_ctl_int %>%
  arrange(- SpatialFDR) %>%
  head() 
da_Myelo_ctl_int %>%
  arrange(- SpatialFDR) %>%
  head() 

da_T_int_end %>%
  arrange(-SpatialFDR) %>%
  head() 

da_Ec_int_end %>%
  arrange(- SpatialFDR) %>%
  head() 
da_Myelo_int_end %>%
  arrange(- SpatialFDR) %>%
  head() 

write.csv(da_T_ctl_int, "SCT/intgr_indiv_SCT_harmony/Subclustering/miloR_DA_sub/T_da_ctlvsint.csv", row.names = FALSE)
write.csv(da_T_int_end, "SCT/intgr_indiv_SCT_harmony/Subclustering/miloR_DA_sub/T_da_endvsint.csv", row.names = FALSE)

write.csv(da_Myelo_ctl_int, "SCT/intgr_indiv_SCT_harmony/Subclustering/miloR_DA_sub/Myelo_da_ctlvsint.csv", row.names = FALSE)
write.csv(da_Myelo_int_end, "SCT/intgr_indiv_SCT_harmony/Subclustering/miloR_DA_sub/Myelo_da_endvsint.csv", row.names = FALSE)

write.csv(da_Ec_ctl_int, "SCT/intgr_indiv_SCT_harmony/Subclustering/miloR_DA_sub/Ec_da_ctlvsint.csv", row.names = FALSE)
write.csv(da_Ec_int_end, "SCT/intgr_indiv_SCT_harmony/Subclustering/miloR_DA_sub/Ec_da_endvsint.csv", row.names = FALSE)
```

```{r}
T_milo <- buildNhoodGraph(T_milo)
Ec_milo <- buildNhoodGraph(Ec_milo)
Myelo_milo <- buildNhoodGraph(Myelo_milo)
```

```{r}
#T_milo$condition <- factor(Myelo$condition, levels = c('ctl', 'int', 'end'))

T1 <- plotUMAP(T_milo, colour_by="cell_subtype", point_size=0.9,text_by = "cell_subtype", text_size = 3.5) + plotNhoodGraphDA(T_milo, da_T_ctl_int, alpha=0.5) +
      plot_layout(guides="auto" )+
      ggtitle("Intermediate vs Control")

T2 <- plotUMAP(T_milo, colour_by="cell_subtype", point_size=0.9, text_by = "cell_subtype", text_size = 3.5) + plotNhoodGraphDA(T_milo, da_T_int_end, alpha=0.5) +
      plot_layout(guides="auto" )+
      ggtitle("Intermediate vs End")
T1
T2

ggsave("SCT/intgr_indiv_SCT_harmony/Subclustering/miloR_DA_sub/Milo_T_ct_to_int.pdf", plot = T1, device = "pdf", dpi = 300, width = 10, height = 7)
ggsave("SCT/intgr_indiv_SCT_harmony/Subclustering/miloR_DA_sub/Milo_T_int_to_end.pdf", plot = T2, device = "pdf", dpi = 300, width = 10, height = 7)
```

```{r}
Ec1 <- plotUMAP(Ec_milo, colour_by="cell_subtype", point_size=0.9,text_by = "cell_subtype", text_size = 3.5) + plotNhoodGraphDA(Ec_milo, da_Ec_ctl_int, alpha=0.5) +
      plot_layout(guides="auto" )+
      ggtitle("Intermediate vs Control")

Ec2 <- plotUMAP(Ec_milo, colour_by="cell_subtype", point_size=0.9, text_by = "cell_subtype", text_size = 3.5) + plotNhoodGraphDA(Ec_milo, da_Ec_int_end, alpha=0.5) +
      plot_layout(guides="auto" )+
      ggtitle("Intermediate vs End")
Ec1
Ec2

ggsave("SCT/intgr_indiv_SCT_harmony/Subclustering/miloR_DA_sub/Milo_Ec_ctl_to_int.pdf", plot = Ec1, device = "pdf", dpi = 300, width = 13, height = 9)
ggsave("SCT/intgr_indiv_SCT_harmony/Subclustering/miloR_DA_sub/Milo_Ec_int_to_end.pdf", plot = Ec2, device = "pdf", dpi = 300, width = 13, height = 9)
```

```{r}
M1 <- plotUMAP(Myelo_milo, colour_by="cell_subtype", point_size=0.9,text_by = "cell_subtype", text_size = 3.5) + plotNhoodGraphDA(Myelo_milo, da_Myelo_ctl_int, alpha=0.5) +
      plot_layout(guides="auto" )+
      ggtitle("Intermediate vs Control")

M2 <- plotUMAP(Myelo_milo, colour_by="cell_subtype", point_size=0.9, text_by = "cell_subtype", text_size = 3.5) + plotNhoodGraphDA(Myelo_milo, da_Myelo_int_end, alpha=0.5) +
      plot_layout(guides="auto" )+
      ggtitle("Intermediate vs End")

ggsave("SCT/intgr_indiv_SCT_harmony/Subclustering/miloR_DA_sub/Milo_Mono_ct_to_int.pdf", plot = M1, device = "pdf", dpi = 300, width = 10, height = 9)
ggsave("SCT/intgr_indiv_SCT_harmony/Subclustering/miloR_DA_sub/Milo_Mono_int_to_end.pdf", plot = M2, device = "pdf", dpi = 300, width = 10, height = 9)
```

```{r}
da_T_ctl_int <- annotateNhoods(T_milo, da_T_ctl_int, coldata_col = "cell_subtype")
da_T_int_end <- annotateNhoods(T_milo, da_T_int_end, coldata_col = "cell_subtype")

da_Ec_ctl_int <- annotateNhoods(Ec_milo, da_Ec_ctl_int, coldata_col = "cell_subtype")
da_Ec_int_end <- annotateNhoods(Ec_milo, da_Ec_int_end, coldata_col = "cell_subtype")

da_Myelo_ctl_int <- annotateNhoods(Myelo_milo, da_Myelo_ctl_int, coldata_col = "cell_subtype")
da_Myelo_int_end <- annotateNhoods(Myelo_milo, da_Myelo_int_end, coldata_col = "cell_subtype")
```

```{r}
# whatever group is less than 0.7 call it mixed
da_T_ctl_int$cell_subtype <- ifelse(da_T_ctl_int$cell_subtype_fraction < 0.7, "Mixed", da_T_ctl_int$cell_subtype)
da_T_int_end$cell_subtype <- ifelse(da_T_int_end$cell_subtype_fraction < 0.7, "Mixed", da_T_int_end$cell_subtype)


da_Ec_ctl_int$cell_subtype <- ifelse(da_Ec_ctl_int$cell_subtype_fraction < 0.7, "Mixed", da_Ec_ctl_int$cell_subtype)
da_Ec_int_end$cell_subtype <- ifelse(da_Ec_int_end$cell_subtype_fraction < 0.7, "Mixed", da_Ec_int_end$cell_subtype)

da_Myelo_ctl_int$cell_subtype <- ifelse(da_Myelo_ctl_int$cell_subtype_fraction < 0.7, "Mixed", da_Myelo_ctl_int$cell_subtype)
da_Myelo_int_end$cell_subtype <- ifelse(da_Myelo_int_end$cell_subtype_fraction < 0.7, "Mixed", da_Myelo_int_end$cell_subtype)

```

```{r}
bee1 <- plotDAbeeswarm(da_T_ctl_int, group.by = "cell_subtype", alpha = 1)+ggtitle("Control to Pretumor")
bee1
ggsave("SCT/intgr_indiv_SCT_harmony/Subclustering/miloR_DA_sub/T_ctl_to_int_beeswarm.pdf", plot = bee1, width = 7, height = 5, dpi = 300)


bee2 <- plotDAbeeswarm(da_T_int_end, group.by = "cell_subtype", alpha = 1)+ggtitle("Pretumor to Tumor")
bee2
ggsave("SCT/intgr_indiv_SCT_harmony/Subclustering/miloR_DA_sub/T_int_to_end_beeswarm.pdf", plot = bee2, width = 7, height = 5, dpi = 300)


bee3 <- plotDAbeeswarm(da_Ec_ctl_int, group.by = "cell_subtype", alpha = 1)+ggtitle("Control to Pretumor")

ggsave("SCT/intgr_indiv_SCT_harmony/Subclustering/miloR_DA_sub/Ec_ctl_to_int_beeswarm.pdf", plot = bee3, width = 7, height = 5, dpi = 300)


bee4 <- plotDAbeeswarm(da_Ec_int_end, group.by = "cell_subtype", alpha = 1)+ggtitle("Pretumor to Tumor")

ggsave("SCT/intgr_indiv_SCT_harmony/Subclustering/miloR_DA_sub/Ec_int_to_end_beeswarm.pdf", plot = bee4, width = 7, height = 5, dpi = 300)

bee5 <- plotDAbeeswarm(da_Myelo_ctl_int, group.by = "cell_subtype", alpha = 1)+ggtitle("Control to Pretumor")
ggsave("SCT/intgr_indiv_SCT_harmony/Subclustering/miloR_DA_sub/Myelo_ctl_to_int_beeswarm.pdf", plot = bee5, width = 7, height = 5, dpi = 300)

bee6 <- plotDAbeeswarm(da_Myelo_int_end, group.by = "cell_subtype", alpha = 1)+ggtitle("Pretumor to Tumor")
ggsave("SCT/intgr_indiv_SCT_harmony/Subclustering/miloR_DA_sub/Myelo_int_to_end_beeswarm.pdf", plot = bee6, width = 7, height = 5, dpi = 300)


```
