==> 6.cellphonedb.Rmd <==
---
title: "Cellphone_db"
output: html_document
date: "2024-04-04"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
seurat_obj <- readRDS(file ="output/Sc_annotated.rds")
```

```{r}
seurat_obj$cell_type_gen <- gsub("_.*", "", seurat_obj$predicted_label)
```

```{r}
if (!dir.exists("outdir")) {
  dir.create("outdir")
}
```

## ortholog transfer-needs debugging

```{r}
if (!requireNamespace("BiocManager", quietly = TRUE))
    install.packages("BiocManager")
BiocManager::install("biomaRt")
```

```{r}

```

```{r}
# Load the biomaRt package needs debug
library(biomaRt)

# Connect to the Ensembl database
ensembl <- useMart("ensembl")


# List available datasets
datasets <- listDatasets(ensembl)

# Print the available datasets
print(datasets)

# Choose the appropriate dataset for mouse (e.g., "mmusculus_gene_ensembl")
mouse_dataset <- "mmusculus_gene_ensembl"

# Use the selected dataset
mouse <- useDataset(mouse_dataset, mart = ensembl)

# Get a list of valid filters for the mouse dataset
mouse_filters <- listFilters(mouse)
print(mouse_filters)

# Based on the list of filters, choose the appropriate filter for mouse Ensembl gene IDs
# For example, if "ensembl_gene_id" is a valid filter, you can use it as follows:
mouse_filter <- "ensembl_gene_id"

# Get the human dataset
human_dataset <- useDataset("hsapiens_gene_ensembl", mart = ensembl)

# Get the mouse gene IDs
mouse_gene_ids <- rownames(seurat_obj)  # Replace with actual mouse gene IDs

# Get the human orthologs
human_orthologs <- getLDS(attributes = c("ensembl_gene_id", "external_gene_name"), 
                           filters = mouse_filter, 
                           values = mouse_gene_ids, 
                           mart = mouse, 
                           attributesL = c("ensembl_gene_id", "external_gene_name"), 
                           martL = human_dataset)
# server id not responding

# Print the human orthologs
print(human_orthologs)


```

```{r}
library(Matrix)
writeMM(obj = seurat_obj@assays$SCT@counts, file = 'outdir/matrix.mtx') # or the normalised counts stored in seurat_obj@assays$RNA@data
# add the features file that i have in the outdir
barcodes <- rownames(seurat_obj@meta.data)
# Optionally, if you want to save these barcodes to a file:
write.table(barcodes, file = "outdir/barcodes.tsv", sep = "\t", quote = FALSE, row.names = FALSE, col.names = FALSE)
```
